used fast_t and least_t in Gregorian functions

+++ Gregorian/Gregorian.h
@@ -4,8 +4,8 @@
 #include "../common/CalendarCache.h"
 #include "../common/YMD.h"
 
-int GregorianEncode(const struct CalendarCache* const restrict cache, struct YMD* const restrict output, const long udn);
+int GregorianEncode(const struct CalendarCache* const restrict cache, struct YMD* const restrict output, const int_fast32_t udn);
 
-int GregorianDecode(const struct CalendarCache* const restrict cache, long* output, const struct YMD* const restrict ymd);
+int GregorianDecode(const struct CalendarCache* const restrict cache, int_fast32_t* output, const struct YMD* const restrict ymd);
 
 #endif
